Stop using the RTC to save on power
Keeping the RTC powered up really increases the amount of power that's
needed when the device is asleep, so let's stop doing it. I don't need
that stuff running for the creature to work right, it was just
nice-to-have stuff for the UI.

Instead of having the creature inject the time, or call out to SNTP to
set the RTC, I can have NodeRED do it on the back end. Rex can draw
whatever power it needs, of course.

This change also calls `esp_sleep_pd_config()` a lot to make the ESP32
shut off as much as it can to save power.

Hopefully with this plus the longer sleep time I'll end up with a device
that uses as little power as I can. If this doesn't add days / weeks to
the power use I'll look at turning off the other circuits (the power
monitor and the amp) via an external circuit, like on a MOSFET.